+++ buildSrc/src/main/kotlin/io/github/androidpoet/supabase/Configuration.kt
@@ -2,7 +2,7 @@ package io.github.androidpoet.supabase
 
 object Configuration {
     const val GROUP = "io.github.androidpoet"
-    const val VERSION = "0.9.2"
+    const val VERSION = "0.9.3"
     const val COMPILE_SDK = 35
     const val MIN_SDK = 21
 }